Refactor filestore tests
