+++ run_tests.py
@@ -22,10 +22,14 @@ test_suites = {
         'AzureJobStoreTest',
         'FileJobStoreTest',
         'GoogleJobStoreTest',
-        'AwsJobStoreCacheTest',
-        'AzureJobStoreCacheTest',
-        'FileJobStoreCacheTest',
-        'GoogleJobStoreCacheTest',
+        'NonCachingFileStoreTestWithFileJobStore',
+        'CachingFileStoreTestWithFileJobStore',
+        'NonCachingFileStoreTestWithAwsJobStore',
+        'CachingFileStoreTestWithAwsJobStore',
+        'NonCachingFileStoreTestWithAzureJobStore',
+        'CachingFileStoreTestWithAzureJobStore',
+        'NonCachingFileStoreTestWithGoogleJobStore',
+        'CachingFileStoreTestWithGoogleJobStore',
         None],
     'integration-test': [
         'CGCloudRNASeqTest',
